+++ definitions/modules/users.module.tsx
@@ -118,7 +118,7 @@ export const USERS_MODULE: Module = {
           value: it => JSX(<jms-e-notes data-id={it.id}/>)
         },
         {
-          value: it => JSX(<jms-e-link link={'/m/users/' + it.id + '/history'}>History</jms-e-link>)
+          value: it => JSX(<jms-e-link icon="history" link={'/m/users/' + it.id + '/history'}>History</jms-e-link>)
         },
         {
           value: it => JSX(<jms-e-tpr data-email={it.data.email}/>),
@@ -276,7 +276,8 @@ export const USERS_MODULE: Module = {
     ],
     subCollections: [
       {name: Collections.History},
-      {name: Collections.Notes}
+      {name: Collections.Notes},
+      {name: 'authorization'}
     ]
   }
 };