+++ functions/copy-definitions.js
@@ -1,4 +1,4 @@
-const {readdir, mkdirSync} = require('fs');
+const {readdir, mkdirSync, cp} = require('fs');
 const { exec } = require('child_process');
 const source = '../definitions';
 const dest = 'definitions';
@@ -14,4 +14,8 @@ readdir(source, (err, files) => {
 			exec(`cp -r "${source}/${file}" "${dest}/${file}"`);
 		}
   });
+
+	exec(`cp definitions/index.functions.ts definitions/index.ts`);
+	exec(`cp definitions/dist/index.functions.d.ts definitions/dist/index.d.ts`);
+	exec(`cp definitions/dist/index.functions.js definitions/dist/index.js`);
 });